refactor: useCopyText
